pyproject.toml
[tool.poetry]
name = "nativas-contracts"
version = "0.1.0"
description = ""
authors = ["Juan Pablo Crespi <juanpablocrespi99@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
eth-brownie = "^1.18.1"
pysha3 = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
